Fix a few binding issues.
To use protocols in API (as parameters or return values), declare an empty
interface with the same name as the protocol prefixed with 'I'
(IBCOVFPSAuthorizationProxy for the BCOVFPSAuthorizationProxy in this case),
and then use the I-prefixed typename as the parameter/return type.

BCOVFPSBrightcoveAuthProxy is not a protocol, it's a class, so remove the
[Protocol, Model] attributes. It also doesn't have a default init selector, so
add the [DisableDefaultCtor] attribute.